+++ example/behaviour_example.dart
@@ -1,18 +1,19 @@
 import 'dart:async';
+import 'dart:developer';
 
 import 'package:behaviour/behaviour.dart';
 
 class CreateCustomer extends Behaviour<CreateCustomerParams, void> {
   @override
   Future<void> action(CreateCustomerParams input, BehaviourTrack? track) {
-    // TODO logic to create a customer
+    log('TODO create customer: ${input.name} - ${input.phoneNumber}');
     return Future.value();
   }
 
   @override
   FutureOr<Exception> onCatchError(
     Object e,
-    StackTrace stacktrace,
+    StackTrace stackTrace,
     BehaviourTrack? track,
   ) {
     return Exception('An unknown error occurred: $e');
@@ -22,12 +23,10 @@ class CreateCustomer extends Behaviour<CreateCustomerParams, void> {
 class CreateCustomerParams {
   const CreateCustomerParams({
     required this.name,
-    required this.address,
     required this.phoneNumber,
   });
 
   final String name;
-  final String address;
   final String phoneNumber;
 }
 
@@ -36,7 +35,6 @@ Future<void> main() async {
   await createCustomer(
     const CreateCustomerParams(
       name: 'Wim',
-      address: 'Somewhere in Belgium',
       phoneNumber: '+32 xxx xx xx xx',
     ),
   );