+++ test/src/behaviour_without_input_test.types.dart
@@ -4,16 +4,15 @@ class _MockAction<TOut> extends Mock implements _Action<TOut> {}
 
 class _BehaviourWithoutInputImpl<TOut> extends BehaviourWithoutInput<TOut> {
   _BehaviourWithoutInputImpl({
-    required Future<TOut> Function(BehaviourTrack? track) action,
+    required FutureOr<TOut> Function(BehaviourTrack? track) action,
     required this.description,
-    BehaviourMonitor? monitor,
-  })  : _action = action,
-        super(monitor: monitor);
+    super.monitor,
+  }) : _action = action;
 
-  final Future<TOut> Function(BehaviourTrack? track) _action;
+  final FutureOr<TOut> Function(BehaviourTrack? track) _action;
 
   @override
-  Future<TOut> action(BehaviourTrack? track) {
+  FutureOr<TOut> action(BehaviourTrack? track) {
     return _action(track);
   }
 
@@ -22,7 +21,13 @@ class _BehaviourWithoutInputImpl<TOut> extends BehaviourWithoutInput<TOut> {
 }
 
 class _Action<TOut> {
-  Future<TOut> action(BehaviourTrack? track) {
+  FutureOr<TOut> action(BehaviourTrack? track) {
     throw UnimplementedError();
   }
+}
+
+class _DummyType {
+  _DummyType(this.value);
+
+  final int value;
 }